backend-api/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Pillow==9.3.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
